Initial commit
At this stage, the files have been moved across from assessory, reorganised slightly, and compile.

Not all the code will work yet, however, as the version of angular.js has been upgraded and some controllers still need to be modified for changes in the Angularjs api.

The URL of the server also needs to be updated.